Renamed file & added themed input component
