+++ src/Tournament/TopBar.tsx
@@ -2,7 +2,7 @@ import { Image, StyleSheet, View } from "react-native";
 import { TransparentCard } from "../Shared";
 import { getImageByType } from "../Shared/APIConverters";
 import PlaceholderImage from "../Shared/PlaceholderImage";
-import { MainText } from "../Shared/ThemedText";
+import { MainText } from "../Shared/ThemedNativeElements";
 import { TopBarProps } from "./types";
 
 export const TopBar = (props: TopBarProps) => {
